Testing out this app with Android wear. Added wear module, and testing out messageAPI to get watch to send command to phone
