+++ LC1/LC1.iml
@@ -89,6 +89,7 @@
     <orderEntry type="library" exported="" name="play-services-6.5.87" level="project" />
     <orderEntry type="library" exported="" name="support-annotations-21.0.3" level="project" />
     <orderEntry type="library" exported="" name="support-v4-21.0.3" level="project" />
+    <orderEntry type="library" name="play-services-wearable-6.5.87" level="project" />
   </component>
 </module>
