+++ LC1/src/main/AndroidManifest.xml
@@ -96,12 +96,15 @@
                 <action android:name="android.intent.action.USER_PRESENT" />
             </intent-filter>
         </service>
-        <service android:name=".DataLayerListenerService" >
+        <service android:name="com.example.harry.wear.DataLayerListenerService" >
             <intent-filter>
-                <action android:name="com.google.android.gms.wearable.BIND_LISTENER" />
+                <action
+                    android:name="com.google.android.gms.wearable.BIND_LISTENER"
+                    android:exported="true"
+                    android:process=":background"
+                    />
             </intent-filter>
         </service>
-
         <receiver android:name=".switchWidget" >
             <intent-filter>
                 <action android:name="android.appwidget.action.APPWIDGET_UPDATE" />
