+++ wear/src/main/java/com/example/harry/wear/MainActivity.java
@@ -3,15 +3,18 @@ package com.example.harry.wear;
 import android.app.Activity;
 import android.os.Bundle;
 import android.support.wearable.view.WatchViewStub;
+import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
 
 import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.common.api.PendingResult;
 import com.google.android.gms.common.api.ResultCallback;
+import com.google.android.gms.wearable.MessageApi;
 import com.google.android.gms.wearable.Node;
 import com.google.android.gms.wearable.NodeApi;
 import com.google.android.gms.wearable.Wearable;
+import com.google.android.gms.wearable.WearableStatusCodes;
 
 import java.util.List;
 
@@ -46,17 +49,20 @@ public class MainActivity extends Activity {
                         final Node node = nodes.get(i);
 
                         // You can just send a message
-                        Wearable.MessageApi.sendMessage(mGoogleApiClient, node.getId(), "/MESSAGE", null);
+                        //Wearable.MessageApi.sendMessage(mGoogleApiClient, node.getId(), "/MESSAGE", null);
 
                         // or you may want to also check check for a result:
-                        // final PendingResult<SendMessageResult> pendingSendMessageResult = Wearable.MessageApi.sendMessage(mGoogleApiClient, node.getId(), "/MESSAGE", null);
-                        // pendingSendMessageResult.setResultCallback(new ResultCallback<MessageApi.SendMessageResult>() {
-                        //      public void onResult(SendMessageResult sendMessageResult) {
-                        //          if (sendMessageResult.getStatus().getStatusCode()==WearableStatusCodes.SUCCESS) {
-                        //              // do something is successed
-                        //          }
-                        //      }
-                        // });
+                        final PendingResult<MessageApi.SendMessageResult> pendingSendMessageResult = Wearable.MessageApi.sendMessage(mGoogleApiClient, node.getId(), "/MESSAGE", null);
+                        pendingSendMessageResult.setResultCallback(new ResultCallback<MessageApi.SendMessageResult>() {
+                              public void onResult(MessageApi.SendMessageResult sendMessageResult) {
+                                  if (sendMessageResult.getStatus().getStatusCode()== WearableStatusCodes.SUCCESS) {
+                                      Log.d("wear", "SUCCESFULLY SENT");
+                                  }
+                                  else{
+                                       Log.d("wear", "Not succesfully sent");
+                                  }
+                              }
+                         });
                     }
                 }
             }
